% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{call_times_two}
\alias{call_times_two}
\title{Multiply a number by two}
\usage{
call_times_two(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Another numeric vector
}
\description{
Multiply a number by two
}
\examples{
call_times_two(3)

}
